pyproject.toml
[project]
name = "rss-feeds-scraper"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Alonso Astroza Tagle", email = "alonsoastroza@gmail.com" }
]
dependencies = [
    "pydantic>=2.6.4",
    "feedparser>=6.0.11",
    "python-dateutil>=2.9.0.post0",
    "beautifulsoup4>=4.12.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/rss_feeds_scraper"]
